[NIXL][P/D] Fix call to tensor.view in TransferEngine
We passed in wrong shape to view when dtype has more than 1 byte

MAX_INTEGRATION_TESTS_REV_ID: 00ac2a841d0fde6a521323d3a6186bf525b000ee

+++ max/tests/integration/pipelines/python/nn/kv_cache/test_transfer_engine.py
@@ -51,10 +51,10 @@ def test_send_recv_basic(device: Device):
     elts_per_page = 3
     num_elts = total_num_pages * elts_per_page
 
-    blocks_1 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int8) + 10).to(
+    blocks_1 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int16) + 10).to(
         device
     )
-    blocks_2 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int8) + 80).to(
+    blocks_2 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int16) + 80).to(
         device
     )
 
@@ -137,10 +137,10 @@ def test_initiate_send_xfer():
     elts_per_page = 3
     num_elts = total_num_pages * elts_per_page
 
-    blocks_1 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int8) + 10).to(
+    blocks_1 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int16) + 10).to(
         device
     )
-    blocks_2 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int8) + 80).to(
+    blocks_2 = Tensor.from_numpy(np.arange(num_elts, dtype=np.int16) + 80).to(
         device
     )
